feature: send percentage as double

+++ service/src/main/java/ch/no1hardy/service/front/dashboard/UserSummary.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class UserSummary {
     private Integer points;
     private Integer ranking;
-    private Integer percentage;
+    private Double percentage;
     private Boolean isConfirmed;
 }
